# docs/api/openapi.frozen.json
{
  "components": {
    "schemas": {
...
      "HRInterviewListResponse": {
        "description": "List payload for HR-facing interview reads.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HRInterviewResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "HRInterviewListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InterviewCancelRequest": {
        "additionalProperties": false,
        "description": "Input payload for staff-triggered interview cancellation.",
        "properties": {
          "cancel_reason_code": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Cancel Reason Code",
            "type": "string"
          }
        },
        "required": [
          "cancel_reason_code"
        ],
        "title": "InterviewCancelRequest",
        "type": "object"
      },
...
      "PublicInterviewActionRequest": {
        "additionalProperties": false,
        "description": "Optional note payload for public candidate interview actions.",
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "PublicInterviewActionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "HRM Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/public/interview-registrations/{token}": {
      "get": {
        "description": "Load current public interview registration payload for one token.",
        "operationId": "get_public_registration_api_v1_public_interview_registrations__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicInterviewRegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Registration",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/public/interview-registrations/{token}/cancel": {
      "post": {
        "description": "Decline interview via public token.",
        "operationId": "cancel_public_registration_api_v1_public_interview_registrations__token__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicInterviewActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicInterviewRegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Public Registration",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/public/interview-registrations/{token}/confirm": {
      "post": {
        "description": "Confirm attendance for current token-bound schedule.",
        "operationId": "confirm_public_registration_api_v1_public_interview_registrations__token__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicInterviewRegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Public Registration",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/public/interview-registrations/{token}/request-reschedule": {
      "post": {
        "description": "Request interview reschedule via public token.",
        "operationId": "request_public_reschedule_api_v1_public_interview_registrations__token__request_reschedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicInterviewActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicInterviewRegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Public Reschedule",
        "tags": [
          "interviews"
        ]
      }
    },
...
    "/api/v1/vacancies/{vacancy_id}/interviews": {
      "get": {
        "description": "List interviews for one vacancy with optional candidate/status filters.",
        "operationId": "list_interviews_api_v1_vacancies__vacancy_id__interviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vacancy Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "candidate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending_sync",
                    "awaiting_candidate_confirmation",
                    "confirmed",
                    "reschedule_requested",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HRInterviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Interviews",
        "tags": [
          "interviews"
        ]
      },
      "post": {
        "description": "Create one active interview and enqueue calendar sync.",
        "operationId": "create_interview_api_v1_vacancies__vacancy_id__interviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vacancy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HRInterviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Interview",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/vacancies/{vacancy_id}/interviews/{interview_id}": {
      "get": {
        "description": "Read one interview by vacancy-scoped identifier.",
        "operationId": "get_interview_api_v1_vacancies__vacancy_id__interviews__interview_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vacancy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HRInterviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Interview",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/vacancies/{vacancy_id}/interviews/{interview_id}/cancel": {
      "post": {
        "description": "Cancel one interview and remove public access immediately.",
        "operationId": "cancel_interview_api_v1_vacancies__vacancy_id__interviews__interview_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vacancy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HRInterviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Interview",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/vacancies/{vacancy_id}/interviews/{interview_id}/feedback": {
      "get": {
        "description": "Read current-version interview feedback summary for HR or assigned interviewer.",
        "operationId": "get_feedback_summary_api_v1_vacancies__vacancy_id__interviews__interview_id__feedback_get",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vacancy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewFeedbackPanelSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Feedback Summary",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/vacancies/{vacancy_id}/interviews/{interview_id}/feedback/me": {
      "put": {
        "description": "Create or replace the current interviewer's feedback for active schedule version.",
        "operationId": "put_feedback_for_current_user_api_v1_vacancies__vacancy_id__interviews__interview_id__feedback_me_put",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vacancy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewFeedbackUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewFeedbackItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Feedback For Current User",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/vacancies/{vacancy_id}/interviews/{interview_id}/reschedule": {
      "post": {
        "description": "Replace schedule window and enqueue new sync.",
        "operationId": "reschedule_interview_api_v1_vacancies__vacancy_id__interviews__interview_id__reschedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vacancy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewRescheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HRInterviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reschedule Interview",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/v1/vacancies/{vacancy_id}/interviews/{interview_id}/resend-invite": {
      "post": {
        "description": "Reissue a fresh public token for the current synchronized schedule.",
        "operationId": "resend_invite_api_v1_vacancies__vacancy_id__interviews__interview_id__resend_invite_post",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vacancy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HRInterviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resend Invite",
        "tags": [
          "interviews"
        ]
      }
    },
...
  }
}